Add privileged mode check for all scripts, build only s390x images for s390x arch (no s390).

+++ ibmjava/tools/cleanup_images.sh
@@ -24,6 +24,18 @@ function usage() {
 args=$1
 [[ "$args" == "" || "$args" == "fclean" ]] || usage
 
+dver=`docker version 2>/dev/null`
+if [ $? != 0 ]; then
+	echo "ERROR: Docker command is running in unprivileged mode."
+	echo "       Run Docker with sudo privileges or make sure the userid is part of the docker group."
+	exit 1
+fi
+
+echo -n "Removing old log files..."
+find .. -name "*.out" -exec rm -f {} \;
+find .. -name "*.err" -exec rm -f {} \;
+echo "done"
+
 echo -n "Removing all containers that have exited..."
 docker ps -q -a | xargs docker rm 2>/dev/null
 echo "done"
@@ -35,6 +47,6 @@ echo "done"
 # Remove all ibmjava images in anticipation for a full build or a pull from remote.
 if [ "$args" == "fclean" ]; then
 	echo -n "Removing all ibmjava images in anticipation for a full build or a pull from remote..."
-	docker rmi -f $(docker images | grep "ibmjava" | awk '{ print $3 }' | uniq) 2>/dev/null
+	docker rmi -f $(docker images | grep -e "ibmjava" -e "j9" | awk '{ print $3 }' | uniq) 2>/dev/null
 	echo "done"
 fi
